requirements.txt
numpy
# opencv-python
pillow
tflite-runtime
flask
mediapipe